requirements.txt
streamlit
gitpython
requests
psycopg2-binary
pandas
sqlglot
openpyxl
boto3>=1.28.0
pytest
flake8
bcrypt>=4.0.1

